requirements.txt
ultralytics==8.1.2
opencv-python==4.9.0.80
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
torch>=2.0.0
tqdm==4.66.1
rich==13.7.0
streamlit==1.31.1
Pillow==10.2.0
wandb==0.16.1
pyyaml==6.0.1
pytest==8.0.0
black==24.1.1
flake8==7.0.0
python-dotenv==1.0.1 